+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   selecting non-transform devices, and does not do so when the `AudioContext`
   is suspended.
 - Generated documentation no longer includes private members.
+- Include the default error message in "meetingStartFailed" and "meetingFailed" events.
 
 ## [2.7.0] - 2021-04-05